Move ApiRequestState to separate file
