+++ frontend/lib/api/index.ts
@@ -3,9 +3,10 @@ export * from "./election";
 export * from "./gen/openapi";
 export * from "./polling_station";
 export * from "./ApiClient";
-export * from "./ApiResult";
 export * from "./ApiProvider";
 export * from "./ApiProviderContext";
+export * from "./ApiRequestState";
+export * from "./ApiResult";
 export * from "./TestUserProvider";
 export * from "./useApiClient";
 export * from "./useApiRequest";
